Add common DateDayMonthYear class to remove duplication in code
