--- document-generator-company-report/src/main/java/uk/gov/companieshouse/document/generator/company/report/mapping/model/document/items/keyfilingdates/items/AccountingReferenceDate.java
@@ -1,31 +1,0 @@
-package uk.gov.companieshouse.document.generator.company.report.mapping.model.document.items.keyfilingdates.items;
-
-import com.fasterxml.jackson.annotation.JsonInclude;
-import com.fasterxml.jackson.annotation.JsonInclude.Include;
-import com.fasterxml.jackson.annotation.JsonProperty;
-
-@JsonInclude(Include.NON_NULL)
-public class AccountingReferenceDate {
-
-    @JsonProperty("day")
-    private String day;
-
-    @JsonProperty("month")
-    private String month;
-
-    public String getDay() {
-        return day;
-    }
-
-    public void setDay(String day) {
-        this.day = day;
-    }
-
-    public String getMonth() {
-        return month;
-    }
-
-    public void setMonth(String month) {
-        this.month = month;
-    }
-}
